# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Forum API",
    "description": "REST API documentation for the forum project. Authentication is handled via a session cookie (`session_token`) set upon login. Endpoints that require authentication are marked with the `cookieAuth` security scheme.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Local development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_token",
        "description": "Session UUID cookie set upon successful login. Has an expiration date. Only one active session per user is allowed at a time."
      }
    },
    "schemas": {
      "PublicUser": {
        "type": "object",
        "description": "Public-facing user information, stripped of sensitive data.",
        "properties": {
          "id": { "type": "integer" },
          "username": { "type": "string" },
          "avatar_path": { "type": "string", "format": "uri-reference" },
          "created_at": { "type": "string", "format": "date-time" }
        },
        "examples": [
          {
            "id": 1,
            "username": "john_doe",
            "avatar_path": "/uploads/avatars/550e8400-e29b-41d4-a716-446655440000.png",
            "created_at": "2024-01-15T10:30:00Z"
          }
        ]
      },
...
      "Response": {
        "type": "object",
        "description": "Standard response wrapper used for all API responses. On success, `data` contains the returned payload. On error, `data` is omitted.",
        "required": ["status", "message"],
        "properties": {
          "status": {
            "type": "string",
            "description": "Indicates whether the request succeeded or failed.",
            "enum": ["success", "error"]
          },
          "message": {
            "type": "string",
            "description": "A human-readable description of the result."
          },
          "data": {
            "description": "The response payload. Present on success; omitted on error."
          }
        },
        "examples": [
          { "status": "error", "message": "A human-readable error message." },
          { "status": "success", "message": "Request completed successfully.", "data": {} }
        ]
      }
    },
    "responses": {
...
      "NotFound": {
        "description": "Not Found - The requested resource does not exist.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Response" },
            "examples": {
              "notFound": {
                "value": { "status": "error", "message": "Resource not found." }
              }
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Internal Server Error - An unexpected error occurred on the server.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Response" },
            "examples": {
              "internalServerError": {
                "value": { "status": "error", "message": "An internal server error occurred." }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Get all users",
        "description": "Returns a list of all registered users. Personal information (e.g. email) is omitted.",
        "responses": {
          "200": {
            "description": "List of all users on the platform.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/Response" },
                    { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/PublicUser" } } } }
                  ]
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
...
    "/users/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Unique ID of the user.",
          "schema": { "type": "integer", "examples": [1] }
        }
      ],
      "get": {
        "tags": ["Users"],
        "summary": "Get a user by ID",
        "description": "Returns public profile information for a specific user. Personal information (e.g. email) is omitted.",
        "responses": {
          "200": {
            "description": "Public details of the requested user.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/Response" },
                    { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/PublicUser" } } }
                  ]
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
...
  }
}